moved loading data related to single admin out of admin home view and into HOC framework
